Extract pane item widget
